+++ core/CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Print inverted question mark for non-printable characters.
+- Remove pre-fill bar from text rendering functions.  [#1173]
 
 ### Deprecated
 
@@ -262,3 +263,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 [#1126]: https://github.com/trezor/trezor-firmware/issues/1126
 [#1139]: https://github.com/trezor/trezor-firmware/issues/1139
 [#1165]: https://github.com/trezor/trezor-firmware/pull/1165
+[#1173]: https://github.com/trezor/trezor-firmware/pull/1173