+++ core/CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Print inverted question mark for non-printable characters.
 - Remove pre-fill bar from text rendering functions.  [#1173]
+- Display coin name when signing or verifying messages.  [#1159]
 
 ### Deprecated
 
@@ -264,5 +265,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 [#1115]: https://github.com/trezor/trezor-firmware/issues/1115
 [#1126]: https://github.com/trezor/trezor-firmware/issues/1126
 [#1139]: https://github.com/trezor/trezor-firmware/issues/1139
+[#1159]: https://github.com/trezor/trezor-firmware/issues/1159
 [#1165]: https://github.com/trezor/trezor-firmware/pull/1165
 [#1173]: https://github.com/trezor/trezor-firmware/pull/1173