+++ core/CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Deprecated
 
 ### Removed
+- ETP, GIN, PTC, ZEL support
 
 ### Fixed
 - CRW addresses are properly generated.  [#1139]